add movement

+++ FlatGalaxySim/FlatGalaxySim/Entities/Position.cs
@@ -8,7 +8,7 @@ namespace FlatGalaxySim.Entities
 {
     public struct Position
     {
-        public int x { get; set; }
-        public int y { get; set; }
+        public double x { get; set; }
+        public double y { get; set; }
     }
 }
